fix upgrade skip not exist index dir

+++ pkg/upgrade/v024.go
@@ -22,6 +22,7 @@ import (
 	"github.com/rs/zerolog/log"
 
 	"github.com/zinclabs/zinc/pkg/config"
+	"github.com/zinclabs/zinc/pkg/zutils"
 )
 
 // UpgradeFromV024 upgrades from version <= 0.2.4
@@ -53,6 +54,9 @@ func UpgradeFromV024() error {
 
 func UpgradeFromV024Index(indexName string) error {
 	rootPath := config.Global.DataPath
+	if ok, _ := zutils.IsExist(path.Join(rootPath, indexName)); !ok {
+		return nil // if index does not exist, skip
+	}
 	if err := os.Rename(path.Join(rootPath, indexName), path.Join(rootPath, indexName+"_old")); err != nil {
 		return err
 	}
